services/worker/pyproject.toml
[project]
name = "worker-service"
version = "0.1.0"
description = "Domain queue worker"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv",
  "boto3",
  "sqlalchemy>=2.0",
  "psycopg[binary]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker"]

